+++ src/main/java/com/jiazhuo/blockgamesquare/controller/TextListController.java
@@ -14,6 +14,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 @Controller
@@ -35,20 +38,20 @@ public class TextListController {
     }
 
     /**
-     * 更新或保存文本
+     * 修改或新增文本
      * @param textList
      * @return
      */
     @RequestMapping(value = "/mgrsite/text/saveOrUpdate", method = RequestMethod.POST)
     @ResponseBody
-    @RequiredPermission("更新或保存文本")
+    @RequiredPermission("修改或新增文本")
     public JSONResultVo saveOrUpdate(TextList textList){
         if (textList.getTid() != null) {
             textListService.update(textList);
-            return JSONResultVo.ok("更新成功");
+            return JSONResultVo.ok("修改成功");
         } else {
             textListService.save(textList);
-            return JSONResultVo.ok("保存成功");
+            return JSONResultVo.ok("新增成功");
         }
     }
 
@@ -57,11 +60,15 @@ public class TextListController {
      * @param tid
      * @return
      */
-    @RequestMapping(value = "/mgrsite/text/delete", method = RequestMethod.DELETE)
+    @RequestMapping(value = "/mgrsite/text/delete", method = RequestMethod.POST)
     @ResponseBody
     @RequiredPermission("删除文本")
     public JSONResultVo delete(Long tid){
         if (tid != null) {
+            TextList textList = textListService.get(tid);
+            if (textList.getType() == 1){
+                return JSONResultVo.error("无法删除");
+            }
             textListService.delete(tid);
             return JSONResultVo.ok("删除成功");
         } else {
@@ -71,19 +78,21 @@ public class TextListController {
 
     /**
      * 获取文本信息(app端)
-     * @param title
-     * @param type
      * @return
      */
     @RequestMapping(value = "/appsite/textList", method = RequestMethod.GET)
     @ResponseBody
-    public SuperResult appText(String title, Byte type){
-        TextList result = textListService.queryText(title, type);
-        SuperResult res = SuperResult.ok(result);
-        if (result == null){
-            return SuperResult.ok("无数据");
+    public JSONResultVo appText(ServletResponse res){
+        HttpServletResponse httpResponse = (HttpServletResponse) res;
+        httpResponse.setHeader("Access-Control-Allow-Origin", "*");
+        httpResponse.setHeader("Access-Control-Allow-Headers","Origin, X-Requested-With, Content-Type, Accept");
+        httpResponse.setHeader("Access-Control-Allow-Credentials", "true");
+        JSONResultVo vo = new JSONResultVo();
+        List<TextList> data = textListService.queryList();
+        if (data == null){
+            return JSONResultVo.error("无数据");
         }
-        res.setMsg(type == 0 ? "获取常见问题成功" : "获取文本信息成功");
-        return res;
+        vo.setResult(data);
+        return vo;
     }
 }
